+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `SwiftUISizingContainer` for handling ideal size and proposed size for a wrapped `UIView`.
 - Added a method to `CollectionViewReorderingDelegate` to check the reordering destination is expected.
 - Added the ability to pass a `CollectionViewConfiguration` to the `CollectionViewController` initializers.
+- Added additional sizing behaviors to `SwiftUIMeasurementContainer` for sizing `UIView`s hosted in 
+  a  SwiftUI `View`.
 
 ### Fixed
 - Fixed sizing of reused `EpoxySwiftUIHostingController`s on iOS 15.2+.